Split classes & add resources
